backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9

